Rename clusters to cluster, since this component is about showing info on a single cluster.
