+++ src/components/cluster/detail/cluster_apps.js
