+++ src/components/cluster/detail/cluster_detail_table.js
