+++ src/components/cluster/detail/index.js
@@ -1,0 +1,28 @@
+'use strict';
+
+import React from 'react';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import ClusterDetailView from './view';
+import { Route, Switch } from 'react-router-dom';
+
+class ClusterDetailIndex extends React.Component {
+  render() {
+    return (
+      <Switch>
+        <Route
+          exact
+          path={`${this.props.match.path}`}
+          component={ClusterDetailView}
+        />
+      </Switch>
+    );
+  }
+}
+
+ClusterDetailIndex.propTypes = {
+  dispatch: PropTypes.func,
+  match: PropTypes.object,
+};
+
+export default connect()(ClusterDetailIndex);
