+++ src/components/cluster/detail/view.js
