+++ src/components/cluster/index.js
@@ -3,26 +3,26 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import ClusterDetailView from './detail/view';
+import ClusterDetailIndex from './detail/';
 import { Route, Switch } from 'react-router-dom';
 
-class ClusterDetailIndex extends React.Component {
+class Clusters extends React.Component {
   render() {
     return (
       <Switch>
         <Route
           exact
           path={`${this.props.match.path}`}
-          component={ClusterDetailView}
+          component={ClusterDetailIndex}
         />
       </Switch>
     );
   }
 }
 
-ClusterDetailIndex.propTypes = {
+Clusters.propTypes = {
   dispatch: PropTypes.func,
   match: PropTypes.object,
 };
 
-export default connect()(ClusterDetailIndex);
+export default connect()(Clusters);
